+++ vedo/shapes.py
@@ -1520,18 +1520,6 @@ def StreamLines(
 
             ![](https://vedo.embl.es/images/volumetric/56964003-9145a500-6b5a-11e9-9d9e-9736d90e1900.png)
     """
-    if len(opts):  # Deprecations
-        printc(" Warning! In StreamLines() unrecognized keywords:", opts, c="y")
-        initial_step_size = opts.pop("initialStepSize", initial_step_size)
-        max_propagation = opts.pop("maxPropagation", max_propagation)
-        max_steps = opts.pop("maxSteps", max_steps)
-        step_length = opts.pop("stepLength", step_length)
-        extrapolate_to_box = opts.pop("extrapolateToBox", extrapolate_to_box)
-        surface_constrained = opts.pop("surfaceConstrained", surface_constrained)
-        compute_vorticity = opts.pop("computeVorticity", compute_vorticity)
-        scalar_range = opts.pop("scalarRange", scalar_range)
-        printc("          Please use 'snake_case' instead of 'camelCase' keywords", c="y")
-
     if isinstance(domain, vedo.Points):
         if extrapolate_to_box:
             grid = _interpolate2vol(domain, **extrapolate_to_box)